Processed the greenpages word doc into a tab seperated file
This is used in the synonyms processing
We should remove the current implementation of synonyms and then
have a new index just for green pages.

The response should list all the matched reponses in the case where
many chatpers match

also we should ignore things like Not to Chapter 89 or
see X unless x is easily found.