Add Green Pages search functionality.
Certain synonyms can be linked to certain Chapters, Headings or
Sections.

E.g. iPad may be in Chapter of Electrical Machinery.

If we know that certain words are linked to one or a few
chapters/headings then return them as possible matches and do not even
do full search in Section,Chapter,Heading,Commodity descriptions.