Add rake task
We want to run our data migration but outside of a deployment, since
it takes hours to run initially. Subsequent runs will essentially be
a no-op (if they aren't, then we are importing invalid data and that's
a bug).

It is expected that this cron task will only be short-lived and removed
in a week or so.