Test all branches on Travis except for release branch
When I develop functionality I'd like Travis to continuously test my
branch before I raise a pull request so that I know when it works or
doesn't. This used to be a problem when we received emails for spec
failures from all branches, but email notifications are now disabled so
it should be desired behaviour.