Measure change feeds does not return proactively added measures
In other words: if I'm querying for commodity changes on 31st of Jan,
I should not be seeing Measures that were added on 24th of Jan but have
validity start date of 1st of Feb. Showing these causes issues with
associations like measure type, because we expect measure type of such
measure to be valid at query date (31st of Jan) and if they were
introduced at the same time that is not the case (also has validity
start date of 1st of Feb).