Rabl caching with memcache in production env
This improves performance of the template generation of the JSON
Implements caching on the headings and commodities
Adds an expiry time for the caches
We need to set the key explicitly as we need the actual_date to
form part of the cache key

use memcache store

either set ENV["MEMCACHE_SERVERS"] if not localhost or change
cache_store's second argument to provide the location of memcache.

Sets global expires_in to 1 day.