Add missing data migration (#58)
There was a data migrations within a schema migrations. This is a bad
practice because we load the structure file when running db:setup.

The file that added this data migration is this:
db/migrate/20130809075350_change_chapter_note_foreign_key_type.rb

This new data migration fixes the issued caused because of this code is
never run.