Fixed issue about a measure missing in the commodity
When a measure was valid at the date selected in TimeMachine but exist a new measure in the future which replaces it, the first measure was not appearing in the results.
The `measures` one_to_many query includes a sub query which contains the measures for base regulations and another subquery which contains the measures for modification regulations: both results are combined and then filtered for valid dates.
But each subquery contained a select unique which filtered the older measures keeping only the last measure of each set. This discarded present queries when future queries existed.

This commit fixes this adding the 'actual' filter condition to each sub-query.
We also wanted to keep the ability to search all the measures when no TimeMachine is being used, so we removed the UNIQUE filter when there is no point in time.