Prevent sending more than 1 email per day
This is a workaround to have a flag that will be switch until the end
of the day.

Currently calling these methods triggers an email, and since the sync
process runs every hour an email is sent every hour.

This change takes advantage of the low level caching system to
invalidate after certain amount of time.

We cannot persist the state of the email sent in the db because none of
this cases are for a specific tariff update.