[MEASURE] Council Regulation - URL generation improvements
**WHAT THIS PR DOES:**

We are introducing conditional council regulation URL generation.
Previously, we used just CELEX number way to generate URL, but
as we noted it doesn't work for all measures.

So that we introduced previously alternative way of URL generation
by Official journal parameters (OJ Seria, OJ number, OJ page and Year).

But after testing we noted that this way doesn't work for old regulations.
For example: R2840/72 and R2658/87
Because they have no official page number.

So in this PR we are using 2 approaches of search for regulation details:

APPROACH 1: If regulation have official journal page
            then we are using seach by Official Journal parameters

APPROACH 2: If regulation doens't have official joirnal page
            then we are using search by celex number

[TRELLO CARD](https://trello.com/c/J4Ddzqgr/453-tariff17-improve-lookup-of-legal-notices)