Update commodity search index to include increased measure details
We've found that we are receiving an elevated number of 500s and many of
them appear to originate from retrieving the overview measures for a
given commodity. To attempt to reduce these we everything required to
display the heading page measures into the elasticsearch index so that
these don't have to be retrieved from the database.

I also used this as a chance to clean up some rubocop violations.