Update commodity serializer to index overview measures
We've found that we are receiving an elevated number of 500s and many of
them appear to originate from retrieving the overview measures for a
given commodity. To attempt to reduce these we everything required to
display the heading page measures into the elasticsearch index so that
these don't have to be retrieved from the database.

This resolves the previous issue in !354 wherein we found multiple VAT
rates with the same description being returned.