Update to disable Rabl caching
We've been finding that we're encountering errors on production that
we're unable to reproduce locally using the same data. After some
investigation we found that the Rabl source cache option seemed to be
the root of these issues and given the other layers of caching we have
present we felt that disabling it was the most prudent decision.