Update manual importer to update the tariff updates table
As part of the new CDS workflow we will be manually importing files, to
avoid importing files that have already been read in and added to the
database we need to update the existing manual importer to both write to
the Tariff Updates table when the imports complete and check if the
update has already been written prior to attempting the import.

We also need to make sure that we don't try to fetch these files from
the S3 bucket in production as we've already downloaded them to the
local file system.